requirements.txt
streamlit
langchain
langchain-community
huggingface_hub
transformers
python-dotenv
pandas
